dom/tests/mochitest/gamepad/test_gamepad_basic.html

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test gamepad  </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">

// Not much we can test here, but this is enough to get the
// gamepad service up and running, which should test that it
// doesn't leak anything.
window.addEventListener("gamepadconnected", function() {});
SimpleTest.ok(true, "dummy check");

</script>
</body>
</html>
